Optimize record updates
At the moment (0.19) it is more efficient to recreate the whole record
`{ field = value }` then to partially update it with `{ record | field = value }`

The latter copies over the fields one by one in a loop, leading to many
small allocations. The former version makes one big allocation and is
therefore a lot faster. This is a detail of the current Kernel
implmentation.